# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Black Tie VoIP Reseller Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: registerReseller
      tags: [auth]
      summary: Register a new reseller (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResellerRegisterRequest"
      responses:
        "200":
          description: Registration submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/resellers:
    get:
      operationId: adminGetResellers
      tags: [admin]
      summary: Get all resellers
      responses:
        "200":
          description: List of resellers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reseller"
    post:
      operationId: adminCreateReseller
      tags: [admin]
      summary: Create reseller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResellerRequest"
      responses:
        "201":
          description: Reseller created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reseller"
  /admin/resellers/{id}:
    get:
      operationId: adminGetReseller
      tags: [admin]
      summary: Get reseller by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reseller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reseller"
    put:
      operationId: adminUpdateReseller
      tags: [admin]
      summary: Update reseller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResellerRequest"
      responses:
        "200":
          description: Reseller updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reseller"
    delete:
      operationId: adminDeleteReseller
      tags: [admin]
      summary: Delete reseller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, reseller]
      required: [email, password, role]
    ResellerRegisterRequest:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
      required: [companyName, contactName, email, password]
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        role:
          type: string
          enum: [admin, reseller]
        user:
          $ref: "#/components/schemas/CurrentUser"
      required: [success, role, user]
    CurrentUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, reseller]
      required: [id, email, name, role]
    Reseller:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        unitStreetNumber:
          type: string
        buildingComplex:
          type: string
        streetName:
          type: string
        address:
          type: string
        address2:
          type: string
        city:
          type: string
        province:
          type: string
        commissionRate:
          type: number
        status:
          type: string
          enum: [active, suspended, pending]
        totalClients:
          type: integer
        monthlyRevenue:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, companyName, contactName, email, status, commissionRate, totalClients, monthlyRevenue, createdAt]
    CreateResellerRequest:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        unitStreetNumber:
          type: string
        buildingComplex:
          type: string
        streetName:
          type: string
        address:
          type: string
        address2:
          type: string
        city:
          type: string
        province:
          type: string
        commissionRate:
          type: number
      required: [companyName, contactName, email, password]
    UpdateResellerRequest:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        unitStreetNumber:
          type: string
        buildingComplex:
          type: string
        streetName:
          type: string
        address:
          type: string
        address2:
          type: string
        city:
          type: string
        province:
          type: string
        commissionRate:
          type: number
        status:
          type: string
          enum: [active, suspended, pending]
...