# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Black Tie VoIP Reseller Portal API
 * OpenAPI spec version: 0.2.0
 */
import type { LoginRequestRole } from "./loginRequestRole";

export interface LoginRequest {
  email: string;
  password: string;
  role: LoginRequestRole;
}